Changed parentSignUp to parentLogins and rerouted
